djs: rename tokenizer-new/ to tokenizer/ now that the old one is gone
Update the three consumer imports (fs/djs/transpiler, fs/djs/parser,
fs/djs/parser/proof) and fix path/prose references across the affected
todo docs. Two unrelated todos (fs/djs/todo/157.md,
fs/bnf/todo/nullable-analysis-shared.md) also mentioned the old path;
157.md's dedup analysis is now marked stale since the code it described
was replaced, rather than silently left pointing at deleted code.
