+++ fs/djs/todo/157.md
@@ -106,8 +106,18 @@ A `serializeValue` factory (in `json/serializer`) parameterized by the extra
 
 ### 3. Tokenizer minus-rewriter
 
-`fs/media/json/tokenizer/module.f.ts:27–101` and `fs/djs/tokenizer/module.f.ts:26–116`
-both wrap `js/tokenizer` and run the same two-state (`'def' | '-'`) `stateScan`
+> Stale as of `fs/djs/tokenizer/todo/replace-old-tokenizer.md`: the old
+> state-machine `fs/djs/tokenizer` this item describes was deleted and
+> replaced by a grammar-based tokenizer (formerly `fs/djs/tokenizer-new`,
+> renamed to `fs/djs/tokenizer`). The same minus-folding pattern was ported
+> into that file's `mapDjsToken`/`parseDjsDefaultState`/`parseDjsMinusState`/
+> `scanDjsToken`, so the dedup opportunity with `fs/media/json/tokenizer`
+> likely still exists, but the line numbers and function names below are
+> outdated — re-verify before acting on this.
+
+`fs/media/json/tokenizer/module.f.ts:27–101` and the old
+`fs/djs/tokenizer/module.f.ts:26–116` (since replaced) both wrapped
+`js/tokenizer` and ran the same two-state (`'def' | '-'`) `stateScan`
 that folds a leading `-` into the following number token:
 
 ```ts
@@ -123,8 +133,11 @@ Both define a matching `ScanState`, a `mapToken` allow-list, `parseDefaultState`
 while DJS keeps them. A `negateOnMinus` scan factory parameterized by the token
 allow-list and an optional bigint case removes the duplication.
 
-> `fs/djs/tokenizer-new/module.f.ts` is a separate, BNF-grammar-based
-> experiment (`todo()`-stubbed) and is **not** part of this duplication.
+> At the time this was written, the BNF-grammar-based tokenizer (then
+> `fs/djs/tokenizer-new/module.f.ts`, `todo()`-stubbed) was a separate
+> experiment and **not** part of this duplication. It has since replaced
+> the tokenizer described above — see the stale-content note at the top of
+> this section.
 
 ### Notes
