+++ fs/djs/transpiler/module.f.ts
@@ -6,7 +6,7 @@
 import { type Unknown } from '../module.f.ts'
 import { type Result, error, ok } from '../../types/result/module.f.ts'
 import { drop, map as listMap, type List, toArray, includes } from '../../types/list/module.f.ts'
-import { tokenize } from '../tokenizer-new/module.f.ts'
+import { tokenize } from '../tokenizer/module.f.ts'
 import { setReplace, at, type OrderedMap } from '../../types/ordered_map/module.f.ts'
 import { stringToList } from '../../text/utf16/module.f.ts'
 import { concat as pathConcat } from '../../path/module.f.ts'
